requirements.txt.txt
Django>=5.0,<6.0
markdown2>=2.4
Pillow>=10.0
django-taggit>=5.0
djangorestframework>=3.15
django-htmx>=1.17
